chore: split up Apply component and disable default export eslint rule
