+++ packages/builder/src/utils/RoundApplicationBuilder.ts
@@ -27,7 +27,7 @@ export default class RoundApplicationBuilder {
     this.ram = ram;
     this.roundAddress = roundAddress;
     this.chainName = chainName;
-
+    debugger;
     const litInit = {
       chain: chainName,
       contract: this.roundAddress,
