+++ packages/builder/src/utils/projects.ts
@@ -4,7 +4,6 @@ import ProjectRegistryABI from "../contracts/abis/ProjectRegistry.json";
 import { getProviderByChainId } from "./utils";
 import { addressesByChainID } from "../contracts/deployments";
 
-// eslint-disable-next-line import/prefer-default-export
 export const fetchProjectOwners = (chainID: ChainId, projectID: string) => {
   const addresses = addressesByChainID(chainID);
   const appProvider = getProviderByChainId(chainID);
